# lawbox-landing/api/api-spec.yml
openapi: 3.0.0
info:
  title: Lawbox Landing API
  description: API Spec
  version: 1.0.0


tags:
  - name: health
    description: Liveness & readiness checks
  - name: leads
    description: Endpoints for capturing and managing prospective customer leads

paths:
  /health:
    get:
      operationId: getAPIHealth
      tags: [health]
      summary: Liveness probe
      description: Returns process status.
      responses:
        '200':
          description: API is alive
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
              examples:
                ok:
                  value:
                    status: ok
        '500':
          description: Unexpected server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /leads:
      post:
        operationId: createLead
        tags: [leads]
        summary: Create a new lead
        description: Accepts email, pain point, and firm size; returns the new lead id or an error.
        requestBody:
          required: true
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NewLead'
              examples:
                example1:
                  value:
                    email: "user@example.com"
                    pain_point: "Document review takes too long"
                    firm_size: "small"
        responses:
          '201':
            description: Lead successfully created
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/CreatedLead'
                examples:
                  success:
                    value:
                      id: 123
                      message: "Lead Successfully Created!"
          '409':
            description: Lead already exists
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/Error'
                examples:
                  duplicate:
                    value:
                      message: "Lead already exists"
          '500':
            description: Database or server error
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/Error'
                examples:
                  failure:
                    value:
                      message: "DB Error Creating Lead"

components:
  schemas:
    HealthResponse:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [ok]
          description: Liveness status
        uptime_s:
          type: number
          format: float
          description: Seconds since process start
    Error:
      type: object
      required: [message]
      properties:
        message:
          type: string
    NewLead:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
        pain_point:
          type: string
          nullable: true
        firm_size:
          type: string
          nullable: true
    CreatedLead:
      type: object
      required: [id, message]
      properties:
        id:
          type: integer
          description: Unique identifier of the lead
        message:
          type: string
          description: Confirmation message